feat: insert games into data layer
as opposed to prerendering the pages based on the games at build time,
put the games in the graphql data layer at build time so gatsby can
handle them naturally and cache them. they will be retrieved at runtime
from the data layer (but not from the third-party sources).

this avoids trying to do fancy stuff around the runtime time zone that's
needed to render the page.

TODO:
- build a resolver out the gql query for a grouping of teams
  - these could potentially reduce some of the ingest logic
- put _all_ the games in the data layer and filter for the upcoming week
  - can the filter logic be wrapped up into a directive
- can a directive handle the timezone?
- make sure the ingest tests still work